requirements.txt
asgiref==3.4.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
idna==2.10
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
python-dotenv==0.15.0
python-firebase==1.2
pytz==2021.3
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.4
